% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrg_fea.R
\name{aggr_fea_voro}
\alias{aggr_fea_voro}
\title{aggr_fea_voro}
\usage{
aggr_fea_voro(nodes, features, type_col)
}
\arguments{
\item{nodes}{a data.frame containing metric x and y coordinates of nodes, additionally an identifier. Coordinates are expected to be the first two columns.}

\item{features}{a data.frame containing metric x and y coordinates of nodes, and feature type. Coordinates are expected to be the first two columns.}

\item{type_col}{a character string naming the columname containing feature types.}
}
\value{
a dataframe with feature types and their corresponding node. Additionally a
    plot is created to display the aggregation of features to nodes. Amount of feature
    per node is added as number to the plot.
}
\description{
Aggregate features by Voronoi tesselation
}
\details{
Creates a Voronoi tesselation of the nodes and aggregates features in tiles according to their spatial location. Returns a dataframe of aggregated features and their associated node as a new collum.
}
\examples{

}
\author{
Oliver Nakoinz <\email{oliver.nakoinz@ufg.uni-kiel.de}>

Chiara Girotto <\email{chiara.girotto@web.de}>

Franziska Faupel <\email{franziska-faupel@gmx.de}>
}
