% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettR_gravity.r
\name{calculate_tij}
\alias{calculate_tij}
\title{Calculate Tij}
\usage{
calculate_tij(Oi, Wj, fcij, alpha)
}
\arguments{
\item{Oi}{Vector of the ouputs of i}

\item{Wj}{Vector of the inputs of j}

\item{fcij}{matrix giving the deterrence function}

\item{alpha}{scaling factor for attractivity}
}
\description{
Calculate Tij
}
\author{
Martin Hinz <martin.hinz@iaw.unibe.ch>
}
