% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecount_fromNodes.R
\name{create_empty_typelist}
\alias{create_empty_typelist}
\title{create empty type list}
\usage{
create_empty_typelist(typelist, aggr_fea)
}
\arguments{
\item{typelist}{create.type.generator}

\item{aggr_fea}{generated nodelist with types as output from arrg_fea.aggr.fea.voro()}
}
\description{
create empty type list
}
\author{
Juliane Watson <<\email{juliane.bonness@web.de}>>
}
