% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_typespectre_sites.R
\name{create_type_generator}
\alias{create_type_generator}
\title{create type generator}
\usage{
create_type_generator(type_tibble, type_col, pre_size)
}
\arguments{
\item{type_tibble}{a tibble with type information (eg. exampledata R1234)}

\item{type_col}{string for column with type information}

\item{pre_size}{amount of letters e.g. charactes before typenumber}
}
\description{
create type generator
}
\examples{

features <- data.frame(x = sample(3433806:3581396, 100, replace = TRUE),
                   y = sample(5286004:5484972, 100, replace = TRUE),
                   type = paste0("B", c(rep(1, 5), rep(2,15), sample(11:19, 20, replace = TRUE), sample(111:119, 30, replace = TRUE), sample(1111:1115, 30, replace = TRUE)))
                   )
                   
create_type_generator(features, "type", 1)
 
}
\author{
Juliane Watson <\email{juliane.bonness@web.de}>
}
