% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecount_fromNodes.R
\name{create_typesectra_for_nodes}
\alias{create_typesectra_for_nodes}
\title{create typespectra for nodes}
\usage{
create_typesectra_for_nodes(node_id, node_type, list_to_modify)
}
\arguments{
\item{node_id}{id of a single node (line[x,2] from output)}

\item{node_type}{type found on a single node (line[x,1] from output)}
}
\description{
create typespectra for nodes
}
\examples{
some example 
}
\author{
Juliane Watson <<\email{juliane.bonness@web.de}>>
}
