% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culdis.R
\name{cul_dist}
\alias{cul_dist}
\title{cul_dist}
\usage{
cul_dist(nodes_x, nodes_y, nodes_id, features, type_col, pre_size = 1,
  method)
}
\arguments{
\item{nodes_x}{a vector containing metric x coordinates of nodes}

\item{nodes_y}{a vector containing metric y coordinates of nodes}

\item{nodes_id}{a vector containing ID for nodes}

\item{features}{a data.frame containing metric x and y coordinates of features, and feature type. Coordinates are expected to be the first two columns.}

\item{type_col}{a character string naming the columname containing feature types.}

\item{pre_size}{numeric, amount of letters, e.g. characters before typenumbers}

\item{method}{character string, the distance measure to be 
used ("euclidean", "maximum", "manhattan", "canberra", 
"binary" or "minkowski")}
}
\value{
bidirectional cultural distance matrix
}
\description{
Cultural distance matrix
}
\details{
Creates a cultural distance matrix by aggregating all features to the nodes by means of a Voronoi tesselation. xxxxxxx. The n dimensional distance matrix of the nodes and their assigned features will be calculated with euclidean distances. The bidirectional matrix will be returned and can be used as network weight.
}
\examples{

set.seed(1234)

cul_dist(nodes_x=sample(3433806:3581396, 10, replace = TRUE),
               nodes_y=sample(5286004:5484972, 10, replace = TRUE), 
               nodes_id=c(1:10), 
             features=data.frame(x=sample(3433806:3581396, 100, replace = TRUE),
                   y=sample(5286004:5484972, 100, replace = TRUE),
                   type=paste0("B", c(rep(1, 5), rep(2,15), sample(11:19, 20, replace = TRUE), sample(111:119, 30, replace = TRUE), sample(1111:1115, 30, replace = TRUE)))
                   ), 
             type_col="type" , pre_size=1, method = "euclidean")
             )

}
\author{
Franziska Faupel <\email{franziska-faupel@gmx.de}>
}
