% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_functions.r
\name{dc2}
\alias{dc2}
\title{Doubly constrained location model (Furness method version)}
\usage{
dc2(Oi, Dj, cij, beta = 1, iterations = 100, detfun = "exp")
}
\arguments{
\item{Oi}{population/workers}

\item{Dj}{settlement size/jobs}

\item{cij}{distance/cost etc. matrix}

\item{beta}{distance decay factor, default = 1}

\item{iterations}{used to stop calculation after n-iterations when
no convergence is achieved}

\item{detfun}{deterrence function (always negative); default is
"power beta"; further option is "exp" for an expontential
function; ..}
}
\value{
a list with the elements:
\itemize{
\item iteration: when was convergence achieved
\item beta: beta (repeated for convenience)
\item Oi: a data.frame showing input and calculated values of summed rows, i.e. sum over j
\item Dj: a data.frame showing input and calculated values of summed columns, i.e. sum over i
\item Ratio: ratio of the difference between targeted and calculated values
\item error: globar error
\item Tij: the resulting flow matrix
\item sumTij: the overall sum of the flow matrix
}
}
\description{
A gravity like approach; the code should be much faster than the dc version; it is based on the Furness method as presented in Ortúzar & Willumsen 2011 184--189
}
\examples{
## From Thomas & Huggett 1980, 150
## --------------------------------------------------
Oi <- c(4,6,2)
Dj <- c(3,8,1)
cij <- matrix(data = c(1,2,2,
                       2,1,2,
                       2,2,1
                       ),
               nr = 3,
               nc = 3
               )
beta <- 1

dc2(Oi = Oi, Dj = Dj, cij = cij, iterations = 5)

## From Ortúzar & Willumsen 2011, 184-189
## --------------------------------------------------
cost_mat <- matrix(data = c(3, 12, 15.5, 24,
                           11, 3, 13, 18,
                           18, 12, 5, 8,
                           22, 19, 7, 5
                           ),
                  nrow = 4,
                  ncol = 4
                  )
Oi_target <- c(400, 460, 400, 702)
Dj_target <- c(260, 400, 500, 802)

dc2(Oi_target, Dj_target, cij = cost_mat, beta = 0.1, detfun = "exp")
}
\references{
Wilson, A.G., Kirkby, M.J., 1980. Mathematics for geographers and planners, 2nd ed, Contemporary problems in geography. Clarendon Pr., Oxford.
Ortúzar S., J. de D., Willumsen, L.G., 2011. Modelling Transport, Fourth edition. ed. John Wiley & Sons, Chichester, West Sussex, United Kingdom.
Thomas, R.W., Huggett, R.J., 1980. Modelling in Geography: A Mathematical Approach. Rowman & Littlefield.
}
