% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_typespectre_sites.R
\name{find_missing_types}
\alias{find_missing_types}
\title{find missing types

Sanity check if each upper level of hierarchical types is represented once in `typelist`.}
\usage{
find_missing_types(type_list, pre_size)
}
\arguments{
\item{pre_size}{amount of letters e.g. charactes before typenumber}

\item{typelist}{a list with type information (eg. exampledata R1234)}
}
\value{
returns flat list of types
}
\description{
find missing types

Sanity check if each upper level of hierarchical types is represented once in `typelist`.
}
\examples{

}
\author{
Juliane Watson <\email{juliane.bonness@web.de}>
}
