% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrg_fea.R
\name{global_bb}
\alias{global_bb}
\title{global_bb}
\usage{
global_bb(nodes, features)
}
\arguments{
\item{nodes}{a data.frame containing metric x and y coordinates of nodes, additionally an identifier. Coordinates are expected to be the first two columns.}

\item{features}{a data.frame containing metric x and y coordinates of nodes, and feature type. Coordinates are expected to be the first two columns.}
}
\value{
an owin object which is slightly bigger, to avoid missing points due to edge effects
}
\description{
Create global bounding box
}
\author{
Oliver Nakoinz <\email{oliver.nakoinz@ufg.uni-kiel.de}>

Chiara Girotto <\email{chiara.girotto@web.de}>

Franziska Faupel <\email{franziska-faupel@gmx.de}>
}
