% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_alonso}
\alias{h_alonso}
\title{h_alonso}
\usage{
h_alonso(E, Si = 1, Sj = 1, f, in_delta, out_delta, in_g,
  in_s = rep(1, length(in_g)), out_g, out_s = rep(1, length(out_g)),
  alpha, x)
}
\arguments{
\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{Si}{a vectory of node attributes, e.g. carrying capacity or population}

\item{Sj}{a vectory of node attributes, e.g. carrying capacity or population}

\item{f}{a scalar giving the general systems constrained}

\item{in_delta}{general node constraining value of inflows}

\item{out_delta}{general node constraining value of outflows}

\item{in_g}{vector of inflows (rows)}

\item{in_s}{vector of inflows (rows)}

\item{out_g}{vector of outflows (rows)}

\item{out_s}{vector of outflows (rows)}
}
\value{
a scalar
}
\description{
Alonso gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
