% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_alpha}
\alias{h_alpha}
\title{h_alpha}
\usage{
h_alpha(alpha, E, f)
}
\arguments{
\item{alpha}{lagrange multiplier}

\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{f}{a scalar giving the general systems constrained}
}
\value{
a scalar
}
\description{
The general flow constrained of the gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
