% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_ariadne}
\alias{h_ariadne}
\title{Hamiltonian ariadne models}
\usage{
h_ariadne(s, v, D, E, k, l, j, u)
}
\arguments{
\item{s}{a vector of node values (e.g. carrying capacity)}

\item{v}{a vector of node values (e.g. exploitation of the carry capacity)}

\item{D}{a matrix of edge deterrence values}

\item{E}{a matrix of edge weight values (e.g. intensity of interaction)}

\item{k}{scalar lagrange multiplier for the kappa term}

\item{l}{scalar lagrange multiplier for the lambda term}

\item{j}{scalar lagrange multiplier for the jay term}

\item{u}{scalar lagrange multiplier for the mu term}
}
\value{
the scalar Hamiltonian value
}
\description{
An implementation of Evans, Rivers & Knappett's \code{ariadne} model; a
Hamiltonian, cost-benefit network model.
}
\examples{

}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
