% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_beta}
\alias{h_beta}
\title{h_beta}
\usage{
h_beta(beta, E, C, c)
}
\arguments{
\item{beta}{langrange multiplier}

\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{C}{distance matrix}

\item{c}{distance constraining cost of the system}
}
\value{
a scalar
}
\description{
The cost constrained of the gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
