% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_constrained_gravity}
\alias{h_constrained_gravity}
\title{h_constrained gravity}
\usage{
h_constrained_gravity(E, Si = 1, Sj = 1, beta, C, c, gammas, g,
  margin = 1)
}
\arguments{
\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{Si}{a vectory of node attributes, e.g. carrying capacity or population}

\item{Sj}{a vectory of node attributes, e.g. carrying capacity or population}

\item{beta}{lagrange multiplier}

\item{C}{distance matrix}

\item{c}{distance constraining cost of the system}

\item{gammas}{lagrange multiplier vector}

\item{g}{constraining vector}

\item{margin}{constraining dimension (either sum of rows or columns)}
}
\value{
a scalar
}
\description{
Constrained gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
