% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_delta}
\alias{h_delta}
\title{h_delta}
\usage{
h_delta(delta, x, g, s)
}
\arguments{
\item{delta}{lagrange multiplier}

\item{x}{general node constraining value}

\item{g}{a vector representing node characteristics}

\item{s}{a vector of node characteristics}
}
\value{
a scalar
}
\description{
The node characteristics constrained of the gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
