% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_double_constrained_gravity}
\alias{h_double_constrained_gravity}
\title{h_double_constrained_gravity}
\usage{
h_double_constrained_gravity(E, Si = 1, Sj = 1, beta, C, c, in_gammas,
  in_g, out_gammas, out_g)
}
\arguments{
\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{Si}{a vectory of node attributes, e.g. carrying capacity or population}

\item{Sj}{a vectory of node attributes, e.g. carrying capacity or population}

\item{beta}{lagrange multiplier}

\item{C}{distance matrix}

\item{c}{distance constraining cost of the system}

\item{in_gammas}{lagrange multiplier vector}

\item{in_g}{vector of inflows (rows)}

\item{out_gammas}{lagrange multiplier vector}

\item{out_g}{vector of outflows (columns)}
}
\value{
a scalar
}
\description{
Double constrained gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
