% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_gamma}
\alias{h_gamma}
\title{h_gamma}
\usage{
h_gamma(gammas, E, g, margin)
}
\arguments{
\item{gammas}{lagrange multiplier vector}

\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{g}{constraining vector}

\item{margin}{constraining dimension (either sum of rows or columns)}
}
\value{
a scalar
}
\description{
The in- or outflow constrained of the gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
