% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_omega}
\alias{h_omega}
\title{h_omega}
\usage{
h_omega(E, Si = 1, Sj = 1)
}
\arguments{
\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{Si}{a vectory of node attributes, e.g. carrying capacity or population}

\item{Sj}{a vectory of node attributes, e.g. carrying capacity or population}
}
\value{
a scalar
}
\description{
Entropy part of the gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
