% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_retail}
\alias{h_retail}
\title{h_retail}
\usage{
h_retail(E, Si = 1, Sj = 1, beta, C, c, gammas, g, delta, x, s)
}
\arguments{
\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{Si}{a vectory of node attributes, e.g. carrying capacity or population}

\item{Sj}{a vectory of node attributes, e.g. carrying capacity or population}

\item{beta}{lagrange multiplier}

\item{C}{distance matrix}

\item{c}{distance constraining cost of the system}

\item{gammas}{lagrange multiplier vector}

\item{g}{constraining vector}

\item{delta}{lagrange multiplier}

\item{x}{general node constraining value}

\item{s}{a vector of node characteristics}
}
\value{
a scalar
}
\description{
Retail Gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
