% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_simple_gravity}
\alias{h_simple_gravity}
\title{h_simple_gravity}
\usage{
h_simple_gravity(E, Si = 1, Sj = 1, f, alpha, beta, C, c)
}
\arguments{
\item{E}{a matrix of random values (to be "optimized" by the Hamiltonian)}

\item{Si}{a vectory of node attributes, e.g. carrying capacity or population}

\item{Sj}{a vectory of node attributes, e.g. carrying capacity or population}

\item{f}{a scalar giving the general systems constrained}

\item{alpha}{lagrange multiplier}

\item{beta}{lagrange multiplier}

\item{C}{distance matrix}

\item{c}{distance constraining cost of the system}
}
\value{
a scalar
}
\description{
Simple gravity model
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
