% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_singly_constrained_gravity_model}
\alias{h_singly_constrained_gravity_model}
\title{Title}
\usage{
h_singly_constrained_gravity_model(alpha, beta, E, c, C, F)
}
\arguments{
\item{alpha}{scalar (random)}

\item{beta}{scalar (random)}

\item{E}{matrix (random 0-1)}

\item{c}{cost matrix}

\item{C}{scalar constant}

\item{F}{scalar constant}
}
\description{
Description
}
