% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{hamiltonian_metrop}
\alias{hamiltonian_metrop}
\title{Title}
\usage{
hamiltonian_metrop(hfunc, hvars, hconsts, beta = 100,
  threshold = 0.001, min_iterations = 50, silent = FALSE)
}
\arguments{
\item{thresholds}{}
}
\description{
Description
}
