% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterrence.R
\name{inverse_exponentional}
\alias{inverse_exponentional}
\title{Inverse exponential function}
\usage{
inverse_exponentional(mat, beta = 0.1)
}
\arguments{
\item{mat}{a distance matrix}

\item{beta}{..}
}
\description{
calculate the cost of distances based on an inverse exponential function
}
\examples{
distmat <- matrix(runif(n=64),8,8)
diag(distmat) <- 0
inverse_exponential(mat = distmat, beta = 1)
}
