% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{inverse_power}
\alias{inverse_power}
\title{Inverse power function}
\usage{
inverse_power(mat, power = 2)
}
\arguments{
\item{mat}{a distance matrix}

\item{power}{the power to be used; default 2}
}
\description{
calculate the cost of distances based on an inverse power law
}
\examples{
distmat <- matrix(runif(n=64),8,8)
diag(distmat) <- 0
inverse_power(mat = distmat, power= 2)
}
