% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{mdm}
\alias{mdm}
\title{Maximum distance model of deterrence}
\usage{
mdm(mat, max)
}
\arguments{
\item{mat}{a distance matrix}

\item{max}{maximum distance of interaction}
}
\value{
a matrix of integer deterrences (0 or 1)

a matrix of integer deterrences (0 or 1)
}
\description{
Creates a deterrence matrix from a distance matrix based on a simple maximum
interaction distance threshold.
}
\examples{
distmat <- matrix(runif(n=64),8,8)
diag(distmat) <- 0
mdm(distmat, 0.1)
}
