% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medieval_settlements}
\alias{medieval_settlements}
\title{Medieval Settlements in Wagrien, Schleswig-Holstein, Germany}
\format{An object of class \code{data.frame} with 175 rows and 4 columns.}
\source{
TODO:add reference
}
\usage{
medieval_settlements
}
\description{
Data digitized from topographic maps
}
\keyword{datasets}
