% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelgraph_functions.R
\name{modelgraph}
\alias{modelgraph}
\title{modelgraph}
\usage{
modelgraph(sfobj, type, par)
}
\arguments{
\item{sfobj}{a sf object containing locations of points}

\item{type}{defining if a Maximum Distance Modell ("mdm") or a Proximal Point Analysis ("ppa") is used to generate the network}

\item{par}{for type "mdm" the threshold value D, and for "ppa" neighbour rank k}
}
\value{
a graph of class "sg" (see package \href{https://cran.r-project.org/web/packages/spatgraphs/index.html}{spatgraphs} for specifics)
}
\description{
Basically a wrapper to produce knn or geometric graphs with spatgraphs.
}
\examples{
set.seed(1985)
d <- data.frame(matrix(runif(15), ncol = 3))
p <- st_as_sf(x = d, coords = 1:2)
result <- modelgraph(sfobj = p, type = "mdm", par = 0.3)
plot(result,st_coordinates(p))
}
