% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelgraph2_functions.R
\name{modelgraph2}
\alias{modelgraph2}
\title{modelgraph2}
\usage{
modelgraph2(sfobj, method = "iom", par = NULL, mode = "undirected")
}
\arguments{
\item{sfobj}{a sf object containing point coordinates}

\item{method}{type of model graph to be produced, either a Maximum Distance Modell ("mdm"), a Proximal Point Analysis ("ppa") or a full graph as base for a Intervening Opportunity Model ("iom"); defaults to "iom" i.e. a full graph.}

\item{par}{for type "mdm" the threshold value D, and for "ppa" neighbour rank k}
}
\value{
a list with
\itemize{
\item{"graph"}{ = a graph object of class tidygraph }
\item{"distance.matrix"}{ = a matrix of all interpoint distances (given in meter)}
\item{"neighbourhood.matrix"}{ = an adjacency (or neighbourhood) matrix, where 0 = not connected nodes and 1 = connected nodes.}
\item{"truncated.distance.matrix"}{ = a matrix, where cell(ij) = distance(ij) if nodes are connected else 0.}
\item{"input.data"}{ = the sf object supplied as input.}
}
}
\description{
Generates model graphs for MDM, PPA or IOM models.
}
\examples{
set.seed(1985)
d <- data.frame(matrix(runif(15), ncol = 3))
p <- sf::st_as_sf(x = d, coords = 1:2)
result <- modelgraph2(sfobj = p, method = "mdm", par = 0.3)
plot(result$graph,sf::st_coordinates(result$input.data))
}
