% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moin_cult_dist.R
\name{moin_cult_dist}
\alias{moin_cult_dist}
\title{moin_cult_dist}
\usage{
moin_cult_dist(nodes_x, nodes_y, nodes_id, features, type_col,
  pre_size = 1, method = "euclidean", plotted = FALSE)
}
\arguments{
\item{nodes_x}{a vector containing metric x coordinates of nodes}

\item{nodes_y}{a vector containing metric y coordinates of nodes}

\item{nodes_id}{a vector containing ID for nodes}

\item{features}{a data.frame containing metric x and y coordinates of features, and feature type. Coordinates are expected to be the first two columns.}

\item{type_col}{a character string naming the columname containing feature types.}

\item{pre_size}{numeric, amount of letters, e.g. characters before typenumbers. Defaults to 1.}

\item{method}{character string, the distance measure to be 
used ("euclidean", "maximum", "manhattan", "canberra", 
"binary" or "minkowski"). Defaults to euclidean distance.}

\item{plotted}{a Boolean operator defining whether a plot should be created. Defaults to FALSE. Edge widths are scaled by maximum distance values and enlarged by factor 2.}
}
\value{
a list containing a graph object of classes tidygraph ("tbl_graph") resp. igraph ("igraph") and the cultural distance matrix.
}
\description{
Cultural distance Network
}
\details{
Creates a network based on a cultural distance matrix
}
\examples{

set.seed(1234)

moin_cult_dist(nodes_x=sample(3433806:3581396, 10, replace = TRUE),
               nodes_y=sample(5286004:5484972, 10, replace = TRUE), 
               nodes_id=c(1:10), 
             features=data.frame(x=sample(3433806:3581396, 100, replace = TRUE),
                   y=sample(5286004:5484972, 100, replace = TRUE),
                   type=paste0("B", c(rep(1, 5), rep(2,15), sample(11:19, 20, replace = TRUE), sample(111:119, 30, replace = TRUE), sample(1111:1115, 30, replace = TRUE)))
                   ), 
             type_col="type" , pre_size=1, method = "euclidean", plotted = TRUE)
             )

}
\author{
Wolfgang Hamer <\email{hamer@geographie.uni-kiel.de}>

Chiara Girotto <\email{chiara.girotto@web.de}>

Hendrik Raese <\email{h.raese@ufg.uni-kiel.de}>

Georg Roth <\email{georg.roth@fu-berlin.de}>
}
