% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moin_cult_network.R
\name{moin_cult_network}
\alias{moin_cult_network}
\title{moin_cult_network}
\usage{
moin_cult_network(nodes_x, nodes_y, nodes_id, features, type_col,
  pre_size = 1, method = "euclidean", network_method = "iom",
  par = NULL, mode = "undirected", crs = NA, plotted = FALSE)
}
\arguments{
\item{nodes_x}{a vector containing metric x coordinates of nodes}

\item{nodes_y}{a vector containing metric y coordinates of nodes}

\item{nodes_id}{a vector containing ID for nodes}

\item{features}{a data.frame containing metric x and y coordinates of features, and feature type. Coordinates are expected to be the first two columns.}

\item{type_col}{a character string naming the columname containing feature types.}

\item{pre_size}{numeric, amount of letters, e.g. characters before typenumbers. Defaults to 1.}

\item{method}{character string, the distance measure to be 
used ("euclidean", "maximum", "manhattan", "canberra", 
"binary" or "minkowski"). Defaults to euclidean distance.}

\item{par}{an integer indicating the distance threshold value D for method "mdm" and the neighbour rank k for "ppa" .}

\item{mode}{a character string; default is undirected. Directed networks are not supported!}

\item{crs}{coordinate reference system: integer with the EPSG code, or character with proj4string.}

\item{plotted}{a Boolean operator defining whether a plot should be created. Defaults to FALSE. Edge widths are scaled by maximum distance values and enlarged by factor 2.}

\item{net_method}{type of model graph to be produced, either a Maximum Distance Model ("mdm"), a Proximal Point Analysis ("ppa") or a full graph as base for a Intervening Opportunity Model ("iom"); defaults to "iom" i.e. a full graph.}

\item{coords_x}{a character string; defines the column name of the X-coordinate in case input is a data.frame; Defaults to NA.}

\item{coords_y}{a character string; defines the column name of the Y-coordinate in case input is a data.frame; Defaults to NA.}
}
\value{
a list containing a graph object of classes tidygraph ("tbl_graph") resp. igraph ("igraph") and the cultural distance matrix.
}
\description{
Creates a network based on the moin_network function and combines the adjacency matrix with the cultural distance matrix.
}
\examples{
# Example mdm network
set.seed(1234)

moin_cult_network(nodes_x=sample(3433806:3581396, 10, replace = TRUE),
                 nodes_y=sample(5286004:5484972, 10, replace = TRUE), 
                 nodes_id=c(1:10), 
                 features=data.frame(x=sample(3433806:3581396, 100, replace = TRUE),
                                     y=sample(5286004:5484972, 100, replace = TRUE),
                                     type=paste0("B", c(rep(1, 5), rep(2,15), sample(11:19, 20, replace = TRUE), 
                                                        sample(111:119, 30, replace = TRUE), sample(1111:1115, 30, replace = TRUE)))
                 ), 
                 type_col="type" , pre_size=1, method = "euclidean", plotted = TRUE,network_method ="mdm", par = 50000, mode = "undirected", crs = NA)

# Example ppa network 
set.seed(1234)
moin_cult_network(nodes_x=sample(3433806:3581396, 10, replace = TRUE),
nodes_y=sample(5286004:5484972, 10, replace = TRUE), 
nodes_id=c(1:10), 
features=data.frame(x=sample(3433806:3581396, 100, replace = TRUE),
                   y=sample(5286004:5484972, 100, replace = TRUE),
                   type=paste0("B", c(rep(1, 5), rep(2,15), sample(11:19, 20, replace = TRUE), 
                                      sample(111:119, 30, replace = TRUE), sample(1111:1115, 30, replace = TRUE)))
), 
type_col="type" , pre_size=1, method = "euclidean", plotted = TRUE,network_method ="ppa", par = 3, mode = "undirected", crs = NA)
}
\author{
Wolfgang Hamer <\email{hamer@geographie.uni-kiel.de}>

Chiara Girotto <\email{chiara.girotto@web.de}>

Hendrik Raese <\email{h.raese@ufg.uni-kiel.de}>

Carolin Tietze <\email{ctietze1991@yahoo.com}>

Michael Bilger <\email{sofnod@googlemail.com}>
}
