% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_typespectre_sites.R
\name{mtypes}
\alias{mtypes}
\title{mtypes

Separating hierarchical type names, so upper levels will be returned. 
  R123 will be returned as R1, R12, R123.}
\usage{
mtypes(type, pre_size)
}
\arguments{
\item{type}{one type from list e.g. R12345}

\item{pre_size}{amount of letters e.g. charactes before typenumber}
}
\value{
returns upper level of type hierarchy as list
}
\description{
mtypes

Separating hierarchical type names, so upper levels will be returned. 
  R123 will be returned as R1, R12, R123.
}
\examples{
returns flat list of meta types

}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
