% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_typespectre_sites.R
\name{parse_all_types}
\alias{parse_all_types}
\title{parse all types

Function which returns a data frame containing type of each feature.}
\usage{
parse_all_types(type_tibble, index_col)
}
\arguments{
\item{type_tibble}{a tibble with type information (eg. exampledata R1234)}

\item{index_col}{index for column to use}
}
\value{
returns a data frame containg all features
}
\description{
parse all types

Function which returns a data frame containing type of each feature.
}
\author{
Juliane Watson <\email{juliane.bonness@web.de}>
}
