% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrg_fea.R
\name{plot_voro_fea}
\alias{plot_voro_fea}
\title{plot_voro_fea}
\usage{
plot_voro_fea(df_aggr_fea, nodes, ppp_nd, voronoi)
}
\arguments{
\item{df_aggr_fea}{a dataframe with feature types and their corresponding node}

\item{nodes}{a data.frame containing metric x and y coordinates of nodes, additionally an identifier. Coordinates are expected to be the first two columns.}

\item{ppp_nd}{a ppp object of nodes}

\item{voronoi}{a tesselation list object generated from nodes}
}
\value{
an owin object which is slightly bigger, to avoid missing points due to edge effects
}
\description{
Plotting Voronoi Tesselation with Amount of Features
}
\author{
Juliane Watson <\email{juliane.bonness@web.de}>
}
