% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettR_gravity.r
\name{rihll_wilson}
\alias{rihll_wilson}
\title{Calculates the Rihll-Wilson-Model}
\usage{
rihll_wilson(Oi, fcij, alpha)
}
\arguments{
\item{Oi}{Outflows originating from i}

\item{fcij}{Deterrence function (has to be calculated from the distance
cost matrix "cij" first)}

\item{alpha}{The scaling factor of attractivity}
}
\value{
a list with the elements:
\itemize{
\item a vector containing Ai
\item a vector containing Ij 
}
}
\description{
Calculates the Rihll-Wilson-Model
}
