% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_functions.r
\name{sc}
\alias{sc}
\title{Singly constrained location model}
\usage{
sc(ei, Pi, Wj, alpha = 1, beta = 1, cij, detfun = "power")
}
\arguments{
\item{ei}{purchasing power, money, etc. of location i}

\item{Pi}{population, size, etc. of location i}

\item{Wj}{attractiveness of location j}

\item{alpha}{default = 1}

\item{beta}{distance decay factor, default = 1}

\item{cij}{distance/cost etc. matrix}

\item{detfun}{deterrence function (always negative); default is
"power beta"; further option is "exp" for an expontential
function (--> entropy maximizing approach; NOTE: beta is
overwritten by the estimate 1/mean(cij); this will be changed
as soon as the beta estimating function is implemented.); ..}
}
\value{
a list with the elements:
\itemize{
\item flows showing the flows from i to j,  
\item si are the sum of the rows, i.e. the sum of i along columns j; this is the factor that can be used to predict, e.g. shopping sales, subject to the constraint of purchasing power (ei) and population (Pi)
\item sj are the constraints
\item K is the proportionality factor
}
}
\description{
A gravity like approach
}
\examples{
## From Wilson & Kirkby 1980, 100f.
ei <- c(2,1,1)
Pi <- c(50, 1000, 500)
eiPi <- ei * Pi
Wj <- c(10, 100, 20)
cij <- matrix(data = c(1, 5, 5,
                      5, 2.585, 5,
                      5, 5, 2),
              nr = 3,
              nc = 3
)

sc(ei = ei, Pi = Pi, Wj = Wj, cij = cij, detfun = "power")
}
\references{
Wilson, A.G., Kirkby, M.J., 1980. Mathematics for geographers and planners, 2nd ed, Contemporary problems in geography. Clarendon Pr., Oxford.
}
