% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettR_gravity.r
\name{simple_gravity}
\alias{simple_gravity}
\title{Simple Gravity Model}
\usage{
simple_gravity(Mi, fcij, k = 1)
}
\arguments{
\item{Mi}{vector of mass for every site}

\item{fcij}{deterrence function from the distance matrix (has to be produced from a distance matrix before)}

\item{k}{adjustment variable}
}
\value{
A matrix with the modeled flows from i to j
}
\description{
Simple Gravity Model
Mass = ability of entities to spread flows = ex : Mass or Population
}
\author{
Clara Filet <clara.filet@gmail.com>
}
