% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.r
\name{spatgraph_transform}
\alias{spatgraph_transform}
\title{spatgraph_transform}
\usage{
spatgraph_transform(x, output = "igraph", mode = "all", ...)
}
\arguments{
\item{x}{a spatgraph object}

\item{output}{a character defining the output as an igraph object ("igraph"), a tidygraph object ("tbl_graph") or a neighbourhood matrix ("matrix")}

\item{mode}{Character scalar, it specifies whether the graph to create is undirected (‘all’ or ‘total’) or directed; and in the latter case, whether it contains the outgoing (‘out’) or the incoming (‘in’) neighbors of the vertices.}

\item{...}{Further parameters given internal to the \code{graph_from_adj_list} function of the igraph package}
}
\value{
depends on the argument output
}
\description{
spatgraph_transform
}
\examples{
x <- matrix(runif(50*2), ncol=2)
g <- spatgraphs::spatgraph(x, "knn", par=3)
#moin::spatgraph_transform(x = g, output = "tidygraph")
# The function can also produce a directed graph 
moin::spatgraph_transform(x = g, output = "tidygraph", mode = "in")

}
