% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettR_gravity.r
\name{deterrence_function}
\alias{deterrence_function}
\title{Calculates the deterrence function}
\usage{
deterrence_function(cij, beta, type = "exponential")
}
\arguments{
\item{cij}{a matrix containing the cost matrix}

\item{beta}{the distance decay factor}

\item{type}{the family of deterrence function to use}
}
\value{
a matrix containing the deterrence function
}
\description{
Calculates the deterrence function used for the different gravity based models
}
\author{
Daniel Knitter <knitter@geographie.uni-kiel.de>

Martin Hinz <martin.hinz@iaw.unibe.ch>

Benedikt Grammer	<benedikt.grammer@univie.ac.at>

Kai Radloff <kai.radloff@hu-berlin.de>

Loren V. Cowin	<Lcowin@gshdl.uni-kiel.de>

Clara Filet <clara.filet@gmail.com>
}
