% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_ariadne}
\alias{h_ariadne}
\title{Title}
\usage{
h_ariadne(s, v, D, E, k, l, j, u)
}
\arguments{
\item{s}{vector}

\item{v}{vector (random 0-1)}

\item{D}{deterrence matrix}

\item{E}{matrix (random 0-1)}

\item{k}{scalar constant}

\item{l}{scalar constant}

\item{j}{scalar constant}

\item{u}{scalar constant}
}
\description{
Description
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
