% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{h_ariadne}
\alias{h_ariadne}
\title{Title}
\usage{
h_ariadne(S, v, d, e, k, l, j, u)
}
\arguments{
\item{S}{vector}

\item{v}{vector (random 0-1)}

\item{d}{deterrence matrix}

\item{e}{matrix (random 0-1)}

\item{k}{scalar constant}

\item{l}{scalar constant}

\item{j}{scalar constant}

\item{u}{scalar constant}
}
\description{
Description
}
