% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{hamiltonian_metrop}
\alias{hamiltonian_metrop}
\title{MCMC/Metropolis algorithm for solving Hamiltonian functions}
\usage{
hamiltonian_metrop(hfunc, hvars, hconsts, hvar_constraints, beta = 100,
  beta_prod = 2, threshold = 0.001, threshold_window = 50,
  silent = FALSE)
}
\arguments{
\item{thresholds}{}
}
\value{

}
\description{
Description
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
