% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{hamiltonian_metrop}
\alias{hamiltonian_metrop}
\title{Title}
\usage{
hamiltonian_metrop(hfunc, hvars, hconsts, beta = 100,
  thresholds = c(0.001, 0.01))
}
\arguments{
\item{thresholds}{}
}
\description{
Title
}
