% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{hamiltonian_metrop}
\alias{hamiltonian_metrop}
\title{MCMC/Metropolis algorithm for solving Hamiltonian functions}
\usage{
hamiltonian_metrop(hfunc, hvars, hconsts, hvar_constraints, beta = 100,
  beta_prod = 2, threshold = 0.001, threshold_window = 50,
  silent = FALSE)
}
\arguments{
\item{hfunc}{a Hamiltonian function either of the Ariadne or one of the gravity models}

\item{hvars}{a list of named arguments to the selected Hamiltonian function}

\item{hconsts}{a list of named constants to the selected Hamiltonian function}

\item{beta}{a scalar; the cooling factor of the "hot Hamiltonian landscape"}

\item{threshold}{a scalar; stopping threshold}

\item{threshold_window}{a scalar; the size of the moving average window calculating the stopping threshold}

\item{silent}{a boolean indicating whether you want to see information of changes in beta and the Hamiltonian per iteration}
}
\value{
a list
}
\description{
Description
}
\author{
Daniel Knitter <\email{knitter@geographie.uni-kiel.de}>

Joe Roe <\email{jwg983@hum.ku.dk}>

Ray Rivers <\email{r.rivers@imperial.ac.uk}>
}
