% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medieval_settlements}
\alias{medieval_settlements}
\title{Medieval Settlements in Wagrien, Schleswig-Holstein, Germany}
\format{A data.frame
\describe{
\item{Name}{The name of the site as written on the map}
\item{rank}{The rank of the site as written on the map}
\item{x}{The x-coordiante in EPSG:25832 as dervied from the georeferenced point}
\item{y}{The y-coordiante in EPSG:25832 as dervied from the georeferenced point}
}}
\source{
TODO:add reference
}
\usage{
medieval_settlements
}
\description{
Data digitized from topographic maps
}
\keyword{datasets}
