% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moin_network.R
\name{moin_network}
\alias{moin_network}
\title{moin_network}
\usage{
moin_network(sfobj, method = "iom", par = NULL, mode = "undirected")
}
\arguments{
\item{sfobj}{a sf object containing point coordinates}

\item{method}{type of model graph to be produced, either a Maximum Distance Modell ("mdm"), a Proximal Point Analysis ("ppa") or a full graph as base for a Intervening Opportunity Model ("iom"); defaults to "iom" i.e. a full graph.}

\item{par}{for type "mdm" the threshold value D, and for "ppa" neighbour rank k}
}
\value{
a list with
\itemize{
\item{"graph"}{ = a graph object of class tidygraph }
\item{"distance_matrix"}{ = a matrix of all interpoint distances (given in meter)}
\item{"neighbourhood_matrix"}{ = an adjacency (or neighbourhood) matrix, where 0 = not connected nodes and 1 = connected nodes.}
\item{"truncated_distance_matrix"}{ = a matrix, where cell(ij) = distance(ij) if nodes are connected else 0.}
\item{"input_data"}{ = the sf object supplied as input.}
}
}
\description{
Generates model graphs for MDM, PPA or IOM models.
}
\examples{
set.seed(1985)
d <- data.frame(matrix(runif(15), ncol = 3))
p <- sf::st_as_sf(x = d, coords = 1:2)
result <- moin_network(sfobj = p, method = "mdm", par = 0.3)
plot(result$graph,sf::st_coordinates(result$input_data))
}
\author{
Hendrik Raese <h.raese@ufg.uni-kiel.de>

Georg Roth <georg.roth@fu-berlin.de>

Wolfgang Hamer <hamer@geographie.uni-kiel.de>
}
