% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_functions.r
\name{sc}
\alias{sc}
\title{Singly constrained location model}
\usage{
sc(Oi, Dj, cij, alpha = 1, beta = 1, detfun = "power")
}
\arguments{
\item{Oi}{origin values, e.g. measured as purchasing power, money, etc. of location i}

\item{Dj}{destination values, e.g. measured as attractiveness of location j}

\item{cij}{distance/cost etc. matrix}

\item{alpha}{default = 1; scaling factor for the attractiveness}

\item{beta}{distance decay factor, default = 1}

\item{detfun}{deterrence function (always negative); default is
"power beta"; further option is "exp" for an expontential
function (--> entropy maximizing approach; NOTE: beta is
overwritten by the estimate 1/mean(cij); this will be changed
as soon as the beta estimating function is implemented.); ..}
}
\value{
a list with the elements:
\itemize{
\item flows showing the flows from i to j,  
\item si are the sum of the rows, i.e. the sum of i along columns j; this is the factor that can be used to predict, e.g. shopping sales, subject to the constraint of purchasing power (ei) and population (Pi)
\item sj are the constraints
\item K is the proportionality factor
}
}
\description{
A gravity like approach
}
\examples{
## From Wilson & Kirkby 1980, 100f.
ei <- c(2,1,1)
Pi <- c(50, 1000, 500)
Wj <- c(10, 100, 20)
cij <- matrix(data = c(1, 5, 5,
                      5, 2.585, 5,
                      5, 5, 2),
              nr = 3,
              nc = 3
)

sc(Oi = ei * Pi, Dj = Wj, cij = cij, detfun = "power")

## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# from: Chan, Y., 2011. Location Theory and Decision
# Analysis. Springer Berlin Heidelberg, Berlin, Heidelberg.
# p. 128f.
cij <- matrix(data = c(8,5,10,
                       3,10,5),                      
              nrow = 2,
              byrow = TRUE)
Vi <- c(1000, 1400)
Vj <- c(1300, 300, 800)

sc(Oi = Vi, Dj = Vj, cij = cij, beta = 2)
}
\references{
Wilson, A.G., Kirkby, M.J., 1980. Mathematics for geographers and planners, 2nd ed, Contemporary problems in geography. Clarendon Pr., Oxford.
}
